+++ libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: ameechan <ameechan@student.42.ch>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/10 15:43:37 by ameechan          #+#    #+#             */
-/*   Updated: 2023/10/19 15:10:55 by ameechan         ###   ########.fr       */
+/*   Updated: 2023/10/19 15:18:17 by ameechan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *dest, int n, size_t len)
 {
 	unsigned char	*ptr;
-	int				i;
+	unsigned long	i;
 
 	i = 0;
 	ptr = (unsigned char *) dest;
